+++ ingredient/ingredient-event-hub-namespace/src/functions.ts
@@ -2,22 +2,25 @@ import {BaseUtility, IngredientManager} from '@azbake/core'
 
 export class EventHubNamespaceUtils extends BaseUtility {
 
-    public get_resource_name(name:string): string {
+    public get_resource_name(shortName: string | null = null): string {
         let util = IngredientManager.getIngredientFunction("coreutils", this.context)
 
         //resource type, name, region enabled
-        const fullName = util.create_resource_name("ehn", name, true);
+        const fullName = util.create_resource_name("ehn", shortName, true);
+
+        this.context._logger.debug(`EventHubNamespaceUtils.get_resource_name() returned ${fullName}`);
+
         return fullName;
     } 
-
-    public async get_resource_group(): Promise<string> {        
+    
+    public async get_resource_profile(shortName: string | null = null, rgShortName: string | null = null): Promise<string> {
         let util = IngredientManager.getIngredientFunction("coreutils", this.context)
+        const name = this.get_resource_name(shortName);
+        const rg = await util.resource_group(rgShortName);
+        const profile = `${rg}/${name}`;
 
-        let override = this.context.Config.rgOverride
-        if (override)
-            return await override.valueAsync(this.context);
-        //resource type, name, region enabled
-        else return util.create_resource_name("ehn", "", true);
+        this.context._logger.debug(`EventHubNamespaceUtils.get_resource_profile() returned ${profile}`);
+        return profile
     }
 }