+++ ingredient/ingredient-metric-alert/src/functions.ts
@@ -1,0 +1,8 @@
+import { BaseUtility, IngredientManager } from '@azbake/core'
+import { ARMHelper } from "@azbake/arm-helper"
+
+export class MetricAlertUtils extends BaseUtility {
+
+
+}
+